BOCM-205 - moves tests for individual appeal details to own file
