+++ apps/api/src/server/app/case-officer/case-officer.get-appeals.test.js
@@ -92,19 +92,3 @@ test('gets the appeals information with received questionnaires', async (t) => {
 	t.deepEqual(resp.body, appealExample);
 });
 
-test('gets the appeals detailed information with received questionnaires', async (t) => {
-	const resp = await request.get('/case-officer/id:');
-	const appealExampleDetail = {
-		AppealId : 1,
-		AppealReference: 'APP/Q9999/D/21/1345264',
-		LocalPlanningDepartment:'Maidstone Borough Council',
-		PlanningApplicationreference:'48269/APP/2021/1482',
-		AppealSiteNearConservationArea: false,
-		WouldDevelopmentAffectSettingOfListedBuilding: false,
-		ListedBuildingDesc: '' // Optional
-	};
-
-	t.is(resp.status, 200);
-	t.deepEqual(resp.body, appealExampleDetail);
-});
-
