+++ apps/api/src/server/app/case-officer/case-officer.routes.js
@@ -1,6 +1,6 @@
 import express from 'express';
 import asyncHandler from '../middleware/async-handler.js';
-import { getAppeals, getAppealsDetail, confirmingLPAQuestionnaire } from './case-officer.controller.js';
+import { getAppeals, getAppealDetails, confirmingLPAQuestionnaire } from './case-officer.controller.js';
 
 const router = express.Router();
 
@@ -13,7 +13,7 @@ router.get('/',
 		}
 	*/
 	getAppeals);
-router.get('/id:', getAppealsDetail);
+router.get('/id:', getAppealDetails);
 router.post('/id:/confirm', asyncHandler(confirmingLPAQuestionnaire));
